requirements.txt
aiohttp==3.6.3
discord.py==1.5.1
python-dotenv==0.15.0
requests==2.24.0
beautifulsoup4==4.9.3
Pillow==8.0.1
numpy==1.19.3
pytz==2020.4
opencv-python==4.4.0.46
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.0
